XXX-0000 Install and use real the config processor
Automate the entire process to checkout the config processor and all
its depenedencies, and install it into a virtual environment.  Update
the code (and auto-generated docs) to reflect that a sucessful run
really only returns a 201 with no other content.

Incorporate tools/setup_env directly into the tox script so that a
you only need to do: tox -e runserver

Change-Id: I16ae249a502cc1aea5246f455c7cb9db0788be96