SCRD-825 Use oslo_config for config processing
Define all config options, along with reasonable production defaults and
help text.  When run without a config file, the service should run
correctly for the day zero installer.
Support the use of oslo-config-generator in tox (tox -e genconfig) in
order to create a sample, documented config file.
Move defaults.cfg to etc/devtest.conf, since it does not need to be
included in the built wheel.  Removed some items from that file that are
unlikely to change.

Change-Id: I2348e32b0bdee6a3b16093e7bba126f98bb68753