XXX-0000 Restore automatic support for local.cfg
When the change to support oslo config was introduced, it did not
retain the ability to automatically load any local.cfg present in the
top-level directory.  The ability to use local.cfg was still present
by using extra options on the tox command, but that was not documented
and little used.  Since oslo config will abort if given a non-existing
filename to load, an empty local.cfg is created if one does not
already exist.

Change-Id: I31bd8ccc1399c41f9555e809fad7e0688e8af5bd