XXX-0000 Exit early from mock tool when requested
The mock_cmd used for mocking out calls to ansible-playbook has a
setting, mock_cmd_exit, that permits it to exit with a non-zero status
in order to simulate errors.   One oddity with this setting was that
it only took effect after all of the mock processing was completed.
In the case of mocking a deployment playbook, the entirety of the
playbook was mocked (including all events and messages), and the exit
code was returned as an afterthought.  Changed the tool to exit early
if the mock_cmd_exit has been set to more accurately simulate a
failure.

Change-Id: I9f73c929f64c4b59321ebd4bec4a1f207133c69f