SCRD-2214 Implement API to restart ardana-service
Create a REST API, /api/v2/restart, that enables the ardana-service to
be restarted after a configurable delay.  The primary use case for
this service is to call it as one of the last steps in the site
playbook in order to forceably restart the service so that it runs
with its updated configuration, which includes keystone
authentication.  The programmable delay enables the service to be
restarted after the site playbook finished; if the service were
immediately restarted while running the site playbook which was in
turn launched from the ardana-service, it it will cause the installer
user interface to hang and it may also cause the site.yml playbook to
abort.

Change-Id: Ic0cd4bce86a282e5d746054bae0d7afeeb315be9