SCRD-0000 Remove filelock
Remove the usage of filelock library around the play status file during
playbook launching and playbook termination.  Playbook launching is
already atomic, so the lock on writing the status file is unnecessary.
Playbook termination is not currently used, but even if it were, there
is no negative ramification if the playbook termination were called
twice in parallel (the worst thing that might happen is that the status
file might be updated twice).

Change-Id: I45412d546cf6860215174626ce65e3ca402c3fdf