SCRD-2049 Reinstate pbr for setup
In commit e944baa, pbr was removed from the setup.py because it was
causing the ardana-service to require the .git repo to be present at
run-time.  But the packaging process now passes in the version number
via the PBR_VERSION environment variable so that the .git repo is not
needed.  Therefore the pbr logic was re-instated, since it yields a
simpler setup and more correct version numbers during development.

Change-Id: Iad450836f700b44b947e1c37b58ca172b8893e62