Update index after commit (bsc#1081376)
Forceably update the index after a commit to work around the fact that
GitPython's commit does not do so.  This is only problematic when new
files are added, since the index still suggests that the files are not
yet tracked, and subsequent invocations of 'git diff-index' will
indicate that there are pending changes.

Change-Id: I4068dbd57e13908d8ae7cee25f0e9e3834bc486d