Permit posting playbook events without a token (bsc#1091740)
This is an adjustment to the fix for bsc#1091740 so that POSTs to the
/api/v2/listener/playbook API do not require a token nor a policy check.
POST calls to that API are used solely by a custom hook for running
ansible playbooks in ardana where it reports on the status of certain
high-level milestones as they are executed in the playbook.  When this
hook is executed, it does not have a keystone auth token, so without
this change the UI will never receive these events.

Unsecuring this API has a couple of foreseeable attacks:
- Posting false positive messages to try to trick the UI into thinking a
  playbook run was successful when it was really not.  This would be
  considered a very low impact.
- A DOS attack (by posting many/large messages to try to fill up the
  disk).

Neither of these are scenarios are likely since:
- Posting to the API requires access to the management network.
- Events can only be posted against a playbook that is running, which
  requires knowing its pid.  This in turn requires either command line
  access to the system or a keystone administrator token; either of
  those scenarios has other *much* greater security ramifications.
- Extra (superfluous) events posted against a running playbook are
  ignored by the UI (their only consumer)

Change-Id: I13cab59e750054b23ec4e76230a8ab52eaa673f1