Set encryption options when encryption key is passed in (bsc#1095885)
The ardana-service is not reading the encryption key from the
location that the UI is passing in. The previous version
of the ardana-service automatically checked for the
encryption-key and reused it as a encrypt and rekey value
when adding a compute host. Mimicking that functionality here
to make progress on fixing the bug

partially-closes: bsc#1095885

Change-Id: I679deb0b1e75ad3c29828a412bc956b2002e9fba