SCRD-3111 Add endpoint for getting internal cp files
Add an endpoint for retrieving the files generated by the config
processor into the "internal" directory.  These files are used by the
topology pages in the day 2 user interface.

As a convenience for developers, during the day 0 validation step, the
several files from the config processor output are copied into
data/cp directory, in a directory structure that mimics production.
These files are then served up by the new api.  Note that this is
only for development: the (default) production configuration already
references the real directories that are populated by the config
processor.

Change-Id: If948c5f83bffc804313727546577e58d0b1ad080