SCRD-3111 Add caching for internal cp files
The yaml files generated in the internal directory by the config
processor tend to be very large, and returning them as json takes many
seconds due to the conversion from yaml to python objects and then
to json.  To improve this turn-around time, the json version of each
file will be cached so that subsequent requests for the same yaml file
can very quickly be returned.

Change-Id: I3de79699c0c72ed390f4700a7593abc29cf84367