Pass encrypt and rekey parms when running config-processor (bsc#1102789)
When running the config-processor during an attempt to add or delete
a compute host, no encrypt or rekey parameters which usually causes
the config processor to prompt for input.  In this case, it throws an
EOFError.  This fix provides the encrypt and rekey parameters as well as
the remove_deleted_servers and free_unused_addresses parameters that
should be set when removing servers.

Change-Id: Iaf144309522e5aea2a1c45290a26413bf28cc560