SCRD-5506 Picked up encprytion key
Currently when UI passes in encyrption key as extraVars with key as
encrypt, ardana-service doesn't pick it up from right var so it misses
to specify the option --vault-password-file when ardana config processor
's data is encrypted. Fixed to pick up the encryption key and
the --vault-password-file option is presented.

Change-Id: I1017aa8cf863682e482e9db2f0945596fac812e1