SCRD-3501 Include IP in cobbler list
The standard cobbler command, 'cobbler system list', returns the system
names as known to cobbler, which may or may not match those used by
ardana, which makes them problematic to use programatically.  Replace
it with a call to 'cobbler system report', which includes much more
information, and parse the output to return the cobbler name and IP
address.
For mock purposes, store the output of the 'cobbler system report'
command from a live system to exercise the production logic for parsing
and returning the output.

Change-Id: Ib89b9d32c866da40606c730c7a0c6118ac39c82a