SCRD-5509 Avoid referencing unassigned variable
If monasca is not in the service catalog (which is mostly in
development/test setups that lack a full complement of services), a
stack trace was being thrown whenever the monasca functions were
executed due to accessing a variable that was never set.  Adjust the
logic to avoid this stack trace.

Change-Id: If87f38892b3b2a542436c5767dad386ff147041f