SCRD-6670 Reuse running ssh-agent or gpg-agent
If there is already an ssh-agent running and connected to the process,
then avoid starting another one.  Since an already-running ssh-agent is
likely to have the user's private key already loaded, this avoids the
extra steps (and failure scenarios) of prompting the user for the
password and loading their private key.

Change-Id: I31aab32cce7ab3f5e0e2affd43a61333e7667bb7