Ignore host with no metrics for checking other host status (SCRD-7803)
When a host is down or the monasca service is not running on it,
the monasca api will still return an entry for that host when
checking other hosts host_alive_status if the "down" host was
at some point an observer of the target host, but the measurements
list will be empty. The code was trying to pull measurements
off of the empty list and having an indexoutofbounds style
exception.

This change checks the list length before attempting to reference
the most recent measurement

Change-Id: I011675dd6eeb1f79cbb2c99643dfbf5895f48665